Update KV value in PDS (#1866)

+++ App/TianHua.Electrical.PDS/Model/ThPDSCircuitModel.cs
@@ -20,7 +20,7 @@
         /// <summary>
         /// 额定电压
         /// </summary>
-        public int KV { get; set; }
+        public double KV { get; set; }
 
         /// <summary>
         /// 相数
